+++ packages/Stun/lib/stun.dart
@@ -2,7 +2,7 @@
 
 library;
 
-export 'src/di/stun_di.dart';
+export 'src/initial_point/initial_point.dart';
 export 'src/implementations/handlers/dual_stun_handler.dart';
 export 'src/implementations/handlers/stun_handler.dart';
 export 'src/implementations/nat/nat_detector.dart';